Add CDC FIFO based on Gray counters
Add the `cdc_fifo_gray` module that implements a clock domain crossing
FIFO queue which synchronizes pointers as Gray coded numbers. Rename the
existing `cdc_fifo` to `cdc_fifo_2phase`. The new implementation has the
advantage of also working if one of the clock domains has stopped; in
such a case the domain that remains active will catch up to the other
and then stop. Adjust the testbench to cover both cases.