Fix tool issues w automatic variables in PLRU tree
This commit moves the declaration of automatic variables to the
beginning of combinatorial blocks and initializes them with a default
value, which avoids issues with Verilator issuing incorrect latch
warnings.  Additionally, it changes the type of `new_index` from integer
to logic to avoid indexing into an integer, which causes issues with
some linting tools.