cdc_4phase: remove synch bypass for DECOUPLED
The previous implementation creates a non-synchronized path
across the CDC, making it significantly more vulnerable
to metastability.

This does not seem to be necessary for DECOUPLED to work in
the cdc_2phase_clearable (cdc_reset_ctrlr).
The important thing seems to be when on the source side the
ready signal is asserted, not how the destination side
synchronizes data.