treewide: Optimizations for faster Verilator simulation (#259)
* rr_arb_tree: Add Verilator pragmas to split tree node signals

This enables Verilator optimizations as it now understands there are no
combinatorial loops. For the Cheshire SoC, this can reduce simulation
time by around 3%.

* lzc: Add Verilator pragmas to split tree node signals

This enables Verilator optimizations as it now understands there are no
combinatorial loops. For the Cheshire SoC, this can reduce simulation
time by around 2.5%.

* lzc: Optimize reversing of input vector for Verilator speedup

In case the vector is not flipped, using a direct assignment can lead to
an approximately 2% decrease in total system simulation time (measured
using the Cheshire SoC).

* cb_filter: Logic simplification for Verilator speed-up

This decreases total simulation time of the Cheshire SoC by around 1.8%.